backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.22
alembic==1.12.1

# NLP and ML
transformers==4.34.1
torch==2.1.0
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.26.0

# Summarization
sentencepiece==0.1.99

# Utilities
python-dotenv==1.0.0
httpx==0.25.0
pytest==7.4.3
